More work implementing QuestPhase Created QuestAreaControl and added it to MainWindow Created QuestAreaViewModel for use by QuestAreaControl

+++ LotR/src/LotR/States/IGame.cs
@@ -19,7 +19,7 @@ namespace LotR.States
     public interface IGame
         : ISource, INotifyPropertyChanged
     {
-        byte CurrentRound { get; }
+        uint CurrentRound { get; }
         IPhase CurrentPhase { get; }
         IQuestArea QuestArea { get; }
         IStagingArea StagingArea { get; }
